refactor(stratum-v2): pool and server on SDK bootstrap; remove legacy IPC client
- Run through SDK module bootstrap and the env contract the node uses for spawned modules
- Update messages, pool, server, template, and library entrypoints accordingly
- Remove the old client and nodeapi IPC layer
- Refresh README, lockfile, and the full stratum test suite

+++ tests/job_cleanup_tests.rs
@@ -13,7 +13,7 @@ fn create_test_block() -> Block {
             timestamp: SystemTime::now()
                 .duration_since(UNIX_EPOCH)
                 .unwrap()
-                .as_secs() as i64,
+                .as_secs(),
             bits: 0x1d00ffff,
             nonce: 0,
         },
